Fix a few more print warnings, add a way to dump the active voices

+++ minibae/BAE_Source/Common/SampleTools.c
@@ -814,7 +814,7 @@ XERR XGetSampleInfoFromSnd(XPTR pResource, SampleDataInfo *pOutInfo)
                         break;
             #endif
                     default:
-                        BAE_PRINTF("Unsupported codec %ld\n", pOutInfo->compressionType);
+                        BAE_PRINTF("Unsupported codec %lu\n", (unsigned long)pOutInfo->compressionType);
                         BAE_ASSERT(FALSE);
                         err = -3;
                         break;
@@ -1037,7 +1037,7 @@ XBYTE               order = X_WORD_ORDER;
             break;
 #endif
         default:
-            BAE_PRINTF("Unsupported codec %ld\n", info->compressionType);
+            BAE_PRINTF("Unsupported codec %lu\n", (unsigned long)info->compressionType);
             BAE_ASSERT(FALSE);
             return NULL;
         }
